+++ internal/js_parser/js_parser.go
@@ -9713,6 +9713,20 @@ func (p *parser) visitExprInOut(expr js_ast.Expr, in exprIn) (js_ast.Expr, exprO
 			e.Args[i] = arg
 		}
 
+		// Warn about calling an import namespace
+		if p.options.outputFormat != config.FormatPreserve {
+			if id, ok := e.Target.Data.(*js_ast.EIdentifier); ok && p.importItemsForNamespace[id.Ref] != nil {
+				r := js_lexer.RangeOfIdentifier(p.source, e.Target.Loc)
+				hint := ""
+				if p.options.ts.Parse {
+					hint = " (make sure to enable TypeScript's \"esModuleInterop\" setting)"
+				}
+				p.log.AddRangeWarning(&p.source, r, fmt.Sprintf(
+					"Cannot call %q because it's an import namespace object, not a function%s",
+					p.symbols[id.Ref.InnerIndex].OriginalName, hint))
+			}
+		}
+
 		// Recognize "require.resolve()" calls
 		if couldBeRequireResolve {
 			if dot, ok := e.Target.Data.(*js_ast.EDot); ok {